+++ recommenders/dnn_sequential_recommender/models/gru4rec.py
@@ -16,7 +16,7 @@ class GRU4Rec(SequentialRecsysModel):
 
     def get_model(self):
         input = layers.Input(shape=(self.max_history_length))
-        x = layers.Embedding(self.num_items + 1, self.embedding_size)(input)
+        x = layers.Embedding(self.num_items + 1, self.embedding_size, dtype='float32')(input)
         for i in range(self.num_gru_layers - 1):
             x = layers.GRU(self.embedding_size, activation=self.activation, return_sequences=True)(x)
         x = layers.GRU(self.embedding_size, activation=self.activation)(x)
